UserService implementation

+++ src/main/java/com/lucas/gym_management/application/service/UserService.java
@@ -1,6 +1,10 @@
 package com.lucas.gym_management.application.service;
 
+import com.lucas.gym_management.application.domain.command.UpdateUserData;
 import com.lucas.gym_management.application.domain.model.*;
+import com.lucas.gym_management.application.domain.model.exceptions.DomainException;
+import com.lucas.gym_management.application.domain.model.valueObjects.Address;
+import com.lucas.gym_management.application.exceptions.BusinessException;
 import com.lucas.gym_management.application.exceptions.NotFoundException;
 import com.lucas.gym_management.application.ports.inbound.create.CreateUserInput;
 import com.lucas.gym_management.application.ports.inbound.create.CreateUserOutput;
@@ -9,22 +13,25 @@ import com.lucas.gym_management.application.ports.inbound.delete.ForDeletingUser
 import com.lucas.gym_management.application.ports.inbound.get.ForGettingUserById;
 import com.lucas.gym_management.application.ports.inbound.get.ForGettingUserByLogin;
 import com.lucas.gym_management.application.ports.inbound.get.GetUserOutput;
+import com.lucas.gym_management.application.ports.inbound.list.ForListingUsers;
+import com.lucas.gym_management.application.ports.inbound.list.ListUserOutput;
 import com.lucas.gym_management.application.ports.inbound.update.ForUpdateUser;
 import com.lucas.gym_management.application.ports.inbound.update.UpdateUserInput;
 import com.lucas.gym_management.application.ports.inbound.update.UpdatedUserOutput;
 import com.lucas.gym_management.application.ports.outbound.repository.UserRepository;
 import jakarta.inject.Named;
 
+import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.UUID;
 
 @Named
 public class UserService implements ForCreatingUser,
         ForGettingUserById,
         ForGettingUserByLogin,
-        ForUpdateUser,
-        ForDeletingUserById {
+        ForListingUsers,
+        ForDeletingUserById,
+        ForUpdateUser {
 
     private final UserRepository userRepository;
 
@@ -35,6 +42,14 @@ public class UserService implements ForCreatingUser,
     @Override
     public CreateUserOutput createUser(final CreateUserInput userInput) {
 
+        if(userRepository.existsByEmail(userInput.email())){
+            throw new BusinessException("Email %s already used.".formatted(userInput.email()));
+        }
+
+        if(userRepository.existsByLogin(userInput.login())){
+            throw new BusinessException("Login %s already used.".formatted(userInput.login()));
+        }
+
         Address userAddress = Address.newAddress(userInput.address().street(),
                 userInput.address().number(),
                 userInput.address().neighborhood(),
@@ -43,46 +58,70 @@ public class UserService implements ForCreatingUser,
                 userInput.address().state());
 
         User newUser = switch(userInput.userType()){
-            case STUDENT -> Student.newStudent(
-                    userInput.name(),
-                    userInput.email(),
-                    userInput.login(),
-                    userInput.password(),
-                    userInput.phone(),
-                    userAddress,
-                    userInput.birthDate(),
-                    userInput.active());
-            case ADMINISTRATOR -> Administrator.newAdministrator(
-                    userInput.name(),
-                    userInput.email(),
-                    userInput.login(),
-                    userInput.password(),
-                    userInput.phone(),
-                    userAddress,
-                    userInput.gymName());
-            case INSTRUCTOR -> Instructor.newInstructor(
-                    userInput.name(),
-                    userInput.email(),
-                    userInput.login(),
-                    userInput.password(),
-                    userInput.phone(),
-                    userAddress,
-                    userInput.cref(),
-                    userInput.specialty());
+            case STUDENT -> {
+                if (userInput.birthDate() == null)
+                    throw new BusinessException("Birth date is required to register a new student.");
+
+                yield Student.newStudent(
+                        userInput.name(),
+                        userInput.email(),
+                        userInput.login(),
+                        userInput.password(),
+                        userInput.phone(),
+                        userAddress,
+                        userInput.birthDate());
+            }
+            case ADMINISTRATOR -> {
+                if (userInput.gymName() == null || userInput.gymName().isBlank())
+                    throw new BusinessException("Gym name is required to register a new administrator.");
+
+                yield Administrator.newAdministrator(
+                        userInput.name(),
+                        userInput.email(),
+                        userInput.login(),
+                        userInput.password(),
+                        userInput.phone(),
+                        userAddress,
+                        userInput.gymName());
+            }
+            case INSTRUCTOR -> {
+                if (userInput.cref() == null || userInput.cref().isBlank() ||
+                        userInput.specialty() == null || userInput.specialty().isBlank())
+                    throw new BusinessException("Cref and specialty are required to register a new instructor.");
+
+                yield Instructor.newInstructor(
+                        userInput.name(),
+                        userInput.email(),
+                        userInput.login(),
+                        userInput.password(),
+                        userInput.phone(),
+                        userAddress,
+                        userInput.cref(),
+                        userInput.specialty());
+            }
         };
 
         var user = userRepository.create(newUser);
+
         return CreateUserOutput.from(user);
+
     }
 
     @Override
     public void deleteUserById(UUID id) {
+        var user = userRepository.findById(id).orElseThrow(
+                () -> new NotFoundException("User with id " + id + " not found")
+        );
+
+        if(user instanceof Student student && student.isActiveMembership())
+            throw new BusinessException("Cannot delete a student with active membership, id %s".formatted(id));
+
         userRepository.deleteById(id);
     }
 
     @Override
     public GetUserOutput getUserById(UUID id) {
-        Optional<User> userById = userRepository.findById(id);
+        var userById = userRepository.findById(id);
 
         return userById.map(GetUserOutput::from)
                 .orElseThrow(()-> new NotFoundException("User with id %s not found".formatted(id)));
@@ -90,11 +129,51 @@ public class UserService implements ForCreatingUser,
 
     @Override
     public GetUserOutput getUserByLogin(String login) {
-        return null;
+        var userByLogin = userRepository.findByLogin(login);
+
+        return userByLogin.map(GetUserOutput::from)
+                .orElseThrow(()-> new NotFoundException("User with login %s not found".formatted(login)));
     }
 
     @Override
-    public UpdatedUserOutput updateUser(UpdateUserInput userInput) {
-        return null;
+    public List<ListUserOutput> listUsers(String name) {
+        List<User> userList = name == null || name.isBlank()
+                ? userRepository.findAll()
+                : userRepository.findByNameLike(name);
+
+        return userList.stream().map(ListUserOutput::from).toList();
+    }
+
+    @Override
+    public UpdatedUserOutput updateUser(UpdateUserInput input) {
+        var userById = userRepository.findById(input.id())
+                .orElseThrow(()-> new NotFoundException("User with id %s not found".formatted(input.id())));
+
+        if(input.email() != null) {
+            if(userRepository.existsByEmailIdNot(input.email(),input.id()))
+                throw new BusinessException("Email %s already used.".formatted(input.email()));
+        }
+
+        Address address = input.address() == null ? null :
+                Address.newAddress(input.address().street(),
+                    input.address().number(),
+                    input.address().neighborhood(),
+                    input.address().zipCode(),
+                    input.address().city(),
+                    input.address().state());
+
+        var updateData = new UpdateUserData(input.name(),
+                input.email(),
+                input.phone(),
+                address,
+                input.gymName(),
+                input.cref(),
+                input.specialty(),
+                input.birthDate(),
+                input.activeMembership());
+
+        userById.applyUpdates(updateData);
+
+        return UpdatedUserOutput.from(userRepository.updateUser(userById));
     }
 }
